backend/requirements.txt
hello-agents[protocols]>=0.2.4,<=0.2.9

fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

httpx>=0.27.0
aiohttp>=3.10.0
openai>=1.109.1

python-dotenv>=1.0.0
python-multipart>=0.0.9
loguru>=0.7.0

fastmcp>=2.0.0
uv>=0.8.0

python-dateutil>=2.8.2
huggingface_hub>=0.25.0
